// challenge/hybris/bin/ext-content/smartedit/web/app/common/services/cache/engine/interfaces/ICacheItem.ts
/*
 * [y] hybris Platform
 *
 * Copyright (c) 2018 SAP SE or an SAP affiliate company. All rights reserved.
 *
 * This software is the confidential and proprietary information of SAP
 * ("Confidential Information"). You shall not disclose such Confidential
 * Information and shall use it only in accordance with the terms of the
 * license agreement you entered into with SAP.
 */
import {Cloneable} from "smarteditcommons/dtos";

export interface ICacheItem<T extends Cloneable> {
	cache: T;
	expirationAge: number;
	evictionTags: string[];
	id: string;
	refreshAge: number;
	timestamp: number;
}